pyproject.toml
[project]
name = "cell-agency"
version = "1.0.0"
description = "Cell Agency — AI Digital Marketing Agency"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.80",
    "fastmcp>=2.0",
    "chromadb>=1.5",
    "anthropic>=0.86",
    "google-generativeai>=0.8",
    "python-telegram-bot>=21.0",
    "httpx>=0.27",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "pillow>=10.0",
    "aiofiles>=23.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.4",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools"]

[tool.ruff]
line-length = 100
target-version = "py311"
